+++ CHANGELOG.md
@@ -9,11 +9,13 @@ and this project is expected to follow [Semantic Versioning](https://semver.org/
 
 ### Added
 
+- Add token unit for shared token data types.
 - Add diagnostics unit for error and diagnostics helpers.
 - Add main compiler program driver.
 
 ### Changed
 
+- Extract token data types from parser into tokens unit.
 - Extract error reporting from compiler into diagnostics unit.
 - Extract main program driver from compiler, convert compiler into a unit.